fix: tutorial view constants

+++ core/src/com/mygdx/tdt4240/states/State.kt
@@ -14,8 +14,7 @@ abstract class State(
     protected val camera: OrthographicCamera = OrthographicCamera()
 
     init {
-        // TODO: change game width and height according to screen size
-        camera.setToOrtho(false, GAME_WIDTH.toFloat(), GAME_HEIGHT.toFloat())
+        camera.setToOrtho(false, GAME_WIDTH, GAME_HEIGHT)
     }
 
     abstract fun update(deltaTime: Float)
